requirements.txt
tensorflow
matplotlib==3.6.1
tqdm
waymo-open-dataset-tf-2-11-0
numpy==1.21.5

